Issue 185 c++11

Changes to allow the gpstk to compile with -std=c++11 option under Solaris Studio 12.4. Closes #185.

See merge request !21

+++ core/lib/GNSSEph/ClockSatStore.cpp
@@ -84,7 +84,7 @@ namespace gpstk
          }
 
          // pull data out of the data table
-         int n,Nlow(Nhalf-1),Nhi(Nhalf),Nmatch(Nhalf);
+         size_t n,Nlow(Nhalf-1),Nhi(Nhalf),Nmatch(Nhalf);
          CommonTime ttag0(it1->first);
          vector<double> times,biases,drifts,accels,sig_biases,sig_drifts,sig_accels;
 
@@ -262,7 +262,7 @@ namespace gpstk
          }
 
          // pull data out of the data table
-         int n,Nhi(Nhalf);
+         size_t n,Nhi(Nhalf);
          CommonTime ttag0(it1->first);
          vector<double> times,biases,drifts;
